+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - PR #1245 Add functions to add pandas and numpy compatibility
 - PR #1260 Add katz_centrality mnmg wrapper
 - PR #1264 CuPy sparse matrix input support for WCC, SCC, SSSP, and BFS
+- PR #1265 Implement Hungarian Algorithm
 
 ## Improvements
 - PR #1227 Pin cmake policies to cmake 3.17 version